Rewrote FPS scripts into ESM
